requirements.txt
fastapi[all]==0.99.1
uvicorn[standard]==0.22.0
tortoise-orm[aiomysql]==0.19.3
aerich==0.7.1
redis==4.5.2
loguru==0.6.0
pycryptodome==3.18.0
aiohttp==3.8.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.1
aiokafka==0.8.1